+++ src/__tests__/yaml.ts
@@ -1,10 +1,10 @@
 import request from 'supertest';
-import { json, ParseErrorHandler } from '..';
+import { ParseErrorHandler, yaml } from '..';
 
-it('should parse request body to a JSON object', async () => {
+it('should parse YAML request body to a JSON object', async () => {
   const app = global.createTestApp();
 
-  app.post('/', json(), async (req, res) => {
+  app.post('/', yaml(), async (req, res) => {
     //            👇 this is no typo 😉
     res.write('typeöf: ' + (typeof req.body) + ' ' + req.body.length);
     res.end();
@@ -14,16 +14,16 @@ it('should parse request body to a JSON object', async () => {
     .post('/')
     .parse(global.parseBody)
     .expect(200)
-    .send({ length: 666 });
+    .send("length: 666\n");
 
   expect(Buffer.isBuffer(response.body)).toBe(true);
   expect(response.body.toString('utf8')).toBe('typeöf: object 666');
 });
 
-it('should parse request body to a JSON object and return 200 if max = 15 and body = 14', async () => {
+it('should parse YAML request body to a JSON object and return 200 if max = 13 and body = 12', async () => {
   const app = global.createTestApp();
 
-  app.post('/', json({ max: 15 }), async (req, res) => {
+  app.post('/', yaml({ max: 13 }), async (req, res) => {
     //            👇 this is no typo 😉
     res.write('typeöf: ' + (typeof req.body) + ' ' + req.body.length);
     res.end();
@@ -33,16 +33,16 @@ it('should parse request body to a JSON object and return 200 if max = 15 and bo
     .post('/')
     .parse(global.parseBody)
     .expect(200)
-    .send(JSON.stringify({ length: 777 }));
+    .send("length: 777\n");
 
   expect(Buffer.isBuffer(response.body)).toBe(true);
   expect(response.body.toString('utf8')).toBe('typeöf: object 777');
 });
 
-it('should parse request body to a JSON object and return 200 if max = 15 and body = 15', async () => {
+it('should parse YAML request body to a JSON object and return 200 if max = 13 and body = 13', async () => {
   const app = global.createTestApp();
 
-  app.post('/', json({ max: 15 }), async (req, res) => {
+  app.post('/', yaml({ max: 13 }), async (req, res) => {
     //            👇 this is no typo 😉
     res.write('typeöf: ' + (typeof req.body) + ' ' + req.body.length);
     res.end();
@@ -52,16 +52,16 @@ it('should parse request body to a JSON object and return 200 if max = 15 and bo
     .post('/')
     .parse(global.parseBody)
     .expect(200)
-    .send(JSON.stringify({ length: 7777 }));
+    .send("length: 7777\n");
 
   expect(Buffer.isBuffer(response.body)).toBe(true);
   expect(response.body.toString('utf8')).toBe('typeöf: object 7777');
 });
 
-it('should parse request body to a JSON object and return 413 if body is bigger than max', async () => {
+it('should parse YAML request body to a JSON object and return 413 if body is bigger than max', async () => {
   const app = global.createTestApp();
 
-  app.post('/', json({ max: 13 }), async (req, res) => {
+  app.post('/', yaml({ max: 11 }), async (req, res) => {
     //            👇 this is no typo 😉
     res.write('typeöf: ' + (typeof req.body) + ' ' + req.body.length);
     res.end();
@@ -71,13 +71,13 @@ it('should parse request body to a JSON object and return 413 if body is bigger 
     .post('/')
     .parse(global.parseBody)
     .expect(413)
-    .send(JSON.stringify({ length: 888 }));
+    .send("length: 888\n");
 });
 
-it('should return 400 on invalid JSON', async () => {
+it('should return 400 on invalid YAML', async () => {
   const app = global.createTestApp();
 
-  app.post('/', json(), async (req, res) => {
+  app.post('/', yaml(), async (req, res) => {
     //            👇 this is no typo 😉
     res.write('typeöf: ' + (typeof req.body) + ' ' + req.body.length);
     res.end();
@@ -87,10 +87,10 @@ it('should return 400 on invalid JSON', async () => {
     .post('/')
     .parse(global.parseBody)
     .expect(400)
-    .send('{ length: 999 ');
+    .send('"length.: ab"c ');
 });
 
-it('should return 406 by custom parse handler on invalid JSON', async () => {
+it('should return 406 by custom parse handler on invalid YAML', async () => {
   const app = global.createTestApp();
 
   const onParseFailed: ParseErrorHandler = async (ctx) => {
@@ -100,7 +100,7 @@ it('should return 406 by custom parse handler on invalid JSON', async () => {
     ));
   };
 
-  app.post('/', json({ onParseFailed }), async (req, res) => {
+  app.post('/', yaml({ onParseFailed }), async (req, res) => {
     //            👇 this is no typo 😉
     res.write('typeöf: ' + (typeof req.body) + ' ' + req.body.length);
     res.end();
@@ -117,5 +117,5 @@ it('should return 406 by custom parse handler on invalid JSON', async () => {
   );
 
   expect(typeof message).toBe('string');
-  expect(message).toBe('SyntaxError: Unexpected token l in JSON at position 2');
+  expect(message).toBe('YAMLException: unexpected end of the stream within a flow collection at line 2, column 1:\n    \n    ^');
 });